Fix setting a default host through command line
When trying to set a default host through the command line, the CLI would report
that `default` is an invalid key. This is because it was not part of the set of
`ConfigOption`'s. The problem with just adding it to there, which I did at first
just to see something working, is that these options are also meant to be
top-level options. I had to redesign the program a little bit to understand the
notion of host-level options and top-level options that can also be host-level
options.

A test is added to ensure that only one host can ever be set as `default`.